Moved LogHabit View and model to intended folder
